Restructure: split handlers/helpers from main, use grate-rs net constants
- Move all syscall handlers from main.rs to handlers.rs
- Move forward_syscall to helpers.rs
- main.rs is now just the entry point with GrateBuilder setup
- Replace local AF_UNIX/AF_INET defs with grate_rs::constants::net
