+++ examples/ipc-grate/src/helpers.rs
@@ -1,0 +1,19 @@
+//! Shared helpers for the IPC grate.
+
+use grate_rs::{getcageid, make_threei_call};
+
+/// Forward a syscall to the next handler via make_threei_call.
+pub fn forward_syscall(
+    nr: u64, calling_cage: u64,
+    args: &[u64; 6], arg_cages: &[u64; 6],
+) -> i32 {
+    let grate_cage = getcageid();
+    match make_threei_call(
+        nr as u32, 0, grate_cage, calling_cage,
+        args[0], arg_cages[0], args[1], arg_cages[1], args[2], arg_cages[2],
+        args[3], arg_cages[3], args[4], arg_cages[4], args[5], arg_cages[5], 0,
+    ) {
+        Ok(r) => r,
+        Err(_) => -1,
+    }
+}
